up final detalhes
