+++ src/components/card/Ca.jsx
@@ -5,13 +5,13 @@ import styles from './ca.module.css'
 const Ca = (x) => {
   return (
     <div>
-      <Card style={{ width: '18rem' , height: '21rem' }} className={styles.ca} >
+      <Card style={{ width: '18rem' , height: '23.5rem' }} className={styles.ca} >
   <Card.Img variant="top" className={styles.cai} src={ x.imagem ?  x.imagem : 'https://ibtaini.com/wp-content/themes/pandastore/assets/images/placeholders/post-placeholder.jpg'} />
   <Card.Body>
     <Card.Title>{x.marca} - {x.modelo}</Card.Title>
     
     
-    <Link className={'btn '+styles.bt} to={'/filmes/detalhe/'+x.id}>{x.nomebotao}{' '}</Link>
+    <Link className={'btn '+styles.bt} to={'busca/'+x.id}>{x.nomebotao}{' '}</Link>
    
   </Card.Body>
 </Card>
