--- src/pages/filmes/Filmes.jsx
@@ -1,121 +1,0 @@
-import React,{useState, useEffect} from 'react'
-import tmdb from '../../services/tmdb'
-import{Row , Col, Button} from "react-bootstrap"
-import Ca from "../../components/card/Ca"
-import styles from "./inicial.module.css"
-import { useNavigate } from 'react-router-dom'
-import { useForm } from 'react-hook-form'
-import {RiSearchEyeLine} from 'react-icons/ri'
-import {AiOutlineArrowDown, AiOutlineArrowUp} from 'react-icons/ai'
-import Swal from 'sweetalert2'
-import logo from './logo.jpeg'
-const Filmes = () => {
-  // declaração de state 
-  const { register, handleSubmit, getValues } = useForm();
-  const navigate = useNavigate
-  function receiveData() {
-    const values = getValues();
-    navigate("/busca", { state: values.busca });
-  }
-  let [filmes,setFilmes] = useState([])
-  const [pages,setPages] = useState()
-  let [page,setPage]=useState()
-  
-  useEffect(()=>{
-    async function data() {
-      const pega = await tmdb.get('posts?_embed&categories=518');
-      console.log(pega.headers["x-wp-totalpages"])
-      setPages(pega.headers["x-wp-totalpages"])
-      const data= pega.data
-      
-      setPage(2)
-      setFilmes(data)
-      
-      
-    }
-    data();
-  },[])
-  
-  function atualizar(){
-    if(page<pages){
-      async function d() {
-        console.log(page);
-        const pega = await tmdb.get('posts?_embed&categories=518&page='+ page);
-        const data= pega.data
-        const nova = filmes.concat(data)
-        console.log(nova);
-        setFilmes(nova)
-        setPage(page+1)
-        console.log(page)
-         console.log(filmes);
-        
-      } 
-      d();
-      }else{
-        Swal.fire({
-          icon: 'error',
-          title: 'Oops...',
-          text: 'Acabaram os posts!',
-          footer: 'Volte em breve'
-        })
-        }
-
-}   
-        
-       
-        
-      
-  return (
-<>
-   <div className={styles.landing}>
-    <div className={styles.titlebox}>
-     <h1 className={styles.title}>APIKI API</h1>
-     <h1 className={styles.title}>Dream system</h1>
-    </div>
-    <div className={styles.rodabox}>
-     <h2 className='text-dark'>Aqui é onde você pode encontrar inúmeras ideias para se atualizar no mundo dos softwares</h2>
-     <h2 className='text-dark'>Um site idealizado para desenvolvedor!</h2>
-    </div>
-    <div className={styles.searchbox}>
-      <h1 className={styles.txt}>Pesquise o assunto que você está procurando aqui</h1>
-      <div className={styles.search}>
-                <form
-                  className="mb-3 "
-                  controlId="busca"
-                  placeholder="Buscar..."
-                >
-                  <input type="text" className={styles.form} {...register("busca")} />
-                </form>
-                <div className={styles.buttonArea}>
-                <Button type="submit" className={"ml-3 "+ styles.btbusca} variant="dark" onClick={handleSubmit(receiveData)} >
-                  <RiSearchEyeLine />
-                    Search Now
-                </Button>
-                </div>
-              </div>
-    </div>
-   </div>
-   <div className={styles.fundol}>
-
-    <div className={styles.listagem} id="lista">
-       <Row xs={1}  xxl={4} xl={4} md={2} sm={1} >
-      
-      {filmes && filmes.map(item=>( 
-        <Col className='mb-3'>   
-         
-          <Ca className={styles.card} id={item.slug} marca={item.title.rendered} modelo={item.status} imagem={ (item._embedded["wp:featuredmedia"]) ? (item._embedded["wp:featuredmedia"][0].source_url) : ("")}  cor={item.link} ano={item.date_gmt} nomebotao='Sabia mais' /> 
-
-        </Col>
-      )) }
-      </Row>
-      {filmes.lenght != 0  && <Button className={styles.bt} onClick={atualizar} ><AiOutlineArrowDown className={styles.icon}/></Button> }
-      
-    </div>
-  </div>
-  
-  <Button className={styles.bt2} href="#lista" ><AiOutlineArrowUp className={styles.icon} /></Button>
-</>
-  )
-}
-
-export default Filmes
