+++ src/pages/filmes/detalhe.module.css
@@ -1,0 +1,379 @@
+.title{
+    padding-top: 1rem;
+
+}
+.subtitle{
+    margin-left: 5rem;
+}
+.cont{
+    width: 100%;
+    padding-left: 10rem;
+    padding-right: 5rem;
+    background:  linear-gradient(0deg, rgba(145,190,212,1) 0%, rgba(217,232,245,1) 50%, rgba(255,255,255,1) 100%);
+}
+.corpo{
+    width: 80%;
+    margin: 0;
+    padding-top: 3rem;
+    padding-left: 5rem;
+    text-align: center;
+    
+}
+.corpo img{
+    width: 50rem;
+    height: 30rem;
+}
+.corpo p{
+    color:black;
+    text-align: left;
+    padding-left: 4rem;
+    font-size: 130%;
+    font-weight: bold;
+}
+.corpo h2{
+    font-size: 200%;
+    padding-top: 1rem;
+    padding-bottom: 1rem;
+}
+.corpo li{
+    font-size: 130%;
+}
+.erro{
+    width: 100%;
+    height: 100%;
+}
+
+/* tabela */
+
+.corpo table {
+    
+    border-left: 5px solid #91BED4;
+    border-right: 5px solid #91BED4;
+    background-color: #D9E8F5;
+    width: 80%;
+    
+    text-align: center;
+    margin-left: 6rem;
+  }
+.corpo table td,.corpo table th {
+    
+    padding: 5px 4px;
+  }
+.corpo table tbody td {
+  
+    font-size: 15px;
+    font-weight: bold;
+    color: black;
+}
+  .corpo table tr:nth-child(even) {
+    
+    background: #91BED4;
+  }
+  .corpo table thead {
+    
+    background: #91BED4;
+  }
+  .corpo table thead th {
+    font-size: 20px;
+    font-weight: bold;
+    color: #FFFFFF;
+    text-align: center;
+  }
+ /* responsive part */
+ @media(max-width:4000px){
+    .title{
+        font-size: 400%;
+    }
+    .subtitle{
+        font-size: 200%;
+    }
+    .corpo p {
+        font-size: 200%;
+    }
+    .corpo h3{
+        font-size: 300%;
+    }
+    .corpo h4{
+        font-size: 300%;
+    }
+    .corpo p a{
+        font-size: 100%;
+    }
+    .corpo a {
+        font-size: 200%;
+    }
+    .corpo li {
+        font-size: 200%;
+        text-align: left;
+    }
+    .corpo h2{
+        font-size: 400%;
+    }
+    .corpo table thead th{
+        font-size: 2rem;
+    }
+    .corpo table tbody td{
+        font-size: 1.5rem;
+        font-weight: bold;
+    }
+    .corpo img{
+        width: 105rem;
+        height: 70rem;
+    }
+    
+    .corpo table{
+        width: 150rem;
+    }
+    .corpo p{
+        font-weight: bold;
+    }
+    .corpo code{
+        font-size: 250%;
+    }
+    .corpo pre{
+        font-size: 250%;
+    }
+    .corpo figcaption{
+        font-size: 160%;
+    }
+ }
+ @media (max-width:2920px){
+    .corpo table{
+        width: 120rem;
+    }
+    .corpo p{
+        font-weight: normal;
+    }
+    .corpo code{
+        font-size: 200%;
+    }
+    .corpo pre{
+        font-size: 200%;
+    }
+    .corpo figcaption{
+        font-size: 140%;
+    }
+ }
+@media(max-width:2350px){
+    .corpo table{
+        width: 92rem;
+    }
+    
+}
+
+@media(max-width:2050px){
+    .corpo p{
+        font-weight: normal;
+    }
+    .corpo table{
+        width: auto;
+    }
+    .cont{
+        padding-right: 0;
+    }
+    .corpo img{
+        width: 70rem;
+        height: 45rem;
+    }
+    .corpo h3{
+        font-size: 200%;
+    }
+    .corpo h4{
+        font-size: 200%;
+    }
+    .corpo p {
+        
+        font-size: 150%;
+    }
+    .corpo p a{
+        font-size: 100%;
+    }
+    .corpo a {
+        
+        font-size: 150%;
+    }
+    .corpo h2{
+        font-size: 250%;
+    }
+    .title{
+        font-size: 250%;
+        
+        
+    }
+    .subtitle{
+        font-size: 130%;
+        
+    }
+    .corpo li{
+        font-size: 150%;
+        text-align: left;
+    }
+    .corpo figcaption{
+        font-size: 120%;
+    }
+}
+@media(max-width:1500px){
+    .corpo img{
+        width: 60rem;
+        height: 40rem;
+    }
+    .corpo code{
+        font-size: 130%;
+    }
+    .corpo pre{
+        font-size: 130%;
+    }
+} 
+@media (max-width:1300px){
+    .cont{
+        text-align: center;
+        padding-left: 5rem;
+        padding-right: 3rem;
+    }
+    .subtitle{
+        margin-left: 21%;
+
+    }
+    .corpo img{
+        width: 45rem;
+        height: 30rem;
+    }
+    .corpo table{
+        width: 100%;
+        margin-left: 2rem;
+    }
+    .corpo p {
+        padding-left: 0;
+        font-size: 120%;
+    }
+    .corpo p a{
+        font-size: 100%;
+    }
+    .corpo a {
+        padding-left: 0;
+        font-size: 120%;
+    }
+    .corpo h2{
+        font-size: 200%;
+    }
+    .title{
+        font-size: 200%;
+        text-align: left;
+        margin-left: 0;
+    }
+    .subtitle{
+        font-size: 100%;
+        text-align: left;
+        margin-left: 0;
+    }
+    .corpo li{
+        font-size: 125%;
+    }
+    .corpo h3{
+        font-size: 180%;
+    }
+    .corpo h4{
+        font-size: 180%;
+    }
+    .corpo figcaption{
+        font-size: 100%;
+    }
+}
+@media (max-width:940px){
+ .corpo img{
+    margin-left: 0;
+    padding-left: 0;
+    width: 40rem;
+    height: 25rem;
+ }
+ 
+}
+@media (max-width:860px){
+    .corpo img {
+        width: 35rem;
+        height: 20rem;
+    }
+    .corpo p{
+        font-weight: normal;
+    }
+}
+@media(max-width:750px){
+    .corpo{
+        margin: 0;
+        padding: 0;
+    }
+    .corpo figcaption{
+        font-size: 80%;
+    }
+}
+@media(max-width:700px){
+    .corpo{
+        text-align: left;
+    }
+    .cont{
+        padding-left: 1rem;
+        padding-right: 0;
+    }
+    .corpo p {
+        padding-left: 0;
+        font-size: 120%;
+    }
+    .corpo p a{
+        font-size: 100%;
+    }
+    .corpo a {
+        padding-left: 0;
+        font-size: 120%;
+    }
+    .corpo h2{
+        font-size: 200%;
+    }
+    .title{
+        font-size: 200%;
+        text-align: left;
+        margin-left: 0;
+    }
+    .subtitle{
+        font-size: 100%;
+        text-align: left;
+        margin-left: 0;
+    }
+    .corpo li{
+        font-size: 130%;
+    }
+    .corpo h3 {
+        font-size: 150%
+    }
+    .corpo h4 {
+        font-size: 150%
+    }
+}
+@media(max-width:600px){
+    .corpo img{
+        width: 30rem;
+        height: 18rem;
+    }
+   
+}
+@media(max-width:520px){
+    .corpo img{
+        width: 100%;
+        height: auto;
+    }
+    .corpo li{
+        font-size: 110%;
+    }
+    
+} 
+@media(max-width;430px){
+    .corpo table{
+        margin-left: 0;
+        
+        width: auto;
+        height: auto;
+    }
+    .corpo div{
+        margin: 0;
+        padding: 0;
+    }
+}
